+++ R/netmeta-internal.R
@@ -15,3 +15,13 @@ function (libname, pkgname)
 
 is.zero <- function(x, n = 10)
   abs(x) < n * .Machine$double.eps
+
+
+invmat <- function(X) {
+  n <- nrow(X)
+  J <- matrix(1, nrow = n, ncol = n)
+  ##
+  res <- solve(X - J / n) + J / n
+  ##
+  res
+}